# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/projects/{project_name}/pipelines/{pipeline_name}/builds": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Get builds",
                "operationId": "many-builds",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Build status (possible values)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Build worker name (exact)",
                        "name": "worker_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of builds",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BuildShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Schedule build",
                "operationId": "create-build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/projects/{project_name}/pipelines/{pipeline_name}/builds/{build_number}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Get the single build",
                "operationId": "single-build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Build number",
                        "name": "build_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Requested build",
                        "schema": {
                            "$ref": "#/definitions/model.Build"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Cancel build",
                "operationId": "update-build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Build number",
                        "name": "build_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated build",
                        "schema": {
                            "$ref": "#/definitions/model.Build"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Build": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildStep"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_name": {
                    "$ref": "#/definitions/sql.NullString"
                }
            }
        },
        "model.BuildLog": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "idx": {
                    "type": "integer"
                },
                "output": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.BuildShort": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_name": {
                    "$ref": "#/definitions/sql.NullString"
                }
            }
        },
        "model.BuildStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "BuildScheduled",
                "BuildRunning",
                "BuildSuccessful",
                "BuildFailed",
                "BuildCanceled"
            ]
        },
        "model.BuildStep": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildLog"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        }
    }
}